the method widens 26 notorch-vlm-2m alloc/copy products to size_t (vendored notorch/vision/stb CodeQL fix)

+++ ariannamethod/notorch.c
@@ -418,7 +418,7 @@ void nt_tape_backward(int loss_idx) {
                 int rows = pw->output->shape[0];
                 int cols = pw->output->ndim >= 2 ? pw->output->shape[1] : pw->output->len / rows;
                 if (rows > 0 && cols > 0) {
-                    float* dw = (float*)calloc(rows * cols, sizeof(float));
+                    float* dw = (float*)calloc((size_t)rows * cols, sizeof(float));
                     if (dw) {
                         for (int i = 0; i < rows; i++)
                             for (int j = 0; j < cols; j++)
@@ -665,7 +665,7 @@ void nt_tape_backward(int loss_idx) {
                 float* gamma_data = NULL;
                 if (has_gamma) gamma_data = g_tape.entries[e->parent2].output->data;
 
-                float* gx = (float*)calloc(T * D, sizeof(float));
+                float* gx = (float*)calloc((size_t)T * D, sizeof(float));
                 float* gg = has_gamma ? (float*)calloc(D, sizeof(float)) : NULL;
                 if (gx) {
                     float* Xrn = px->output->data;
@@ -711,9 +711,9 @@ void nt_tape_backward(int loss_idx) {
                 int T = (int)e->aux;
                 int D = (int)e->aux2;
                 float sc = 1.0f / sqrtf((float)D);
-                float* dq = (float*)calloc(T * D, sizeof(float));
-                float* dk = (float*)calloc(T * D, sizeof(float));
-                float* dv = (float*)calloc(T * D, sizeof(float));
+                float* dq = (float*)calloc((size_t)T * D, sizeof(float));
+                float* dk = (float*)calloc((size_t)T * D, sizeof(float));
+                float* dv = (float*)calloc((size_t)T * D, sizeof(float));
                 if (dq && dk && dv) {
                     for (int i = 0; i < T; i++) {
                         float* qi = pq->output->data + i * D;
@@ -774,9 +774,9 @@ void nt_tape_backward(int loss_idx) {
                 int D = e->output->len / T;
                 int n_heads = D / head_dim;
                 float sc = 1.0f / sqrtf((float)head_dim);
-                float* dq = (float*)calloc(T * D, sizeof(float));
-                float* dk = (float*)calloc(T * D, sizeof(float));
-                float* dv = (float*)calloc(T * D, sizeof(float));
+                float* dq = (float*)calloc((size_t)T * D, sizeof(float));
+                float* dk = (float*)calloc((size_t)T * D, sizeof(float));
+                float* dv = (float*)calloc((size_t)T * D, sizeof(float));
                 if (dq && dk && dv) {
                     for (int h = 0; h < n_heads; h++) {
                         int ho = h * head_dim;
@@ -843,9 +843,9 @@ void nt_tape_backward(int loss_idx) {
                 int KV_D = n_kv_heads * head_dim;
                 int gqa_ratio = n_heads / n_kv_heads;
                 float sc = 1.0f / sqrtf((float)head_dim);
-                float* dq = (float*)calloc(T * Q_D, sizeof(float));
-                float* dk = (float*)calloc(T * KV_D, sizeof(float));
-                float* dv = (float*)calloc(T * KV_D, sizeof(float));
+                float* dq = (float*)calloc((size_t)T * Q_D, sizeof(float));
+                float* dk = (float*)calloc((size_t)T * KV_D, sizeof(float));
+                float* dv = (float*)calloc((size_t)T * KV_D, sizeof(float));
                 if (dq && dk && dv) {
                     for (int h = 0; h < n_heads; h++) {
                         int kv_h = h / gqa_ratio;
@@ -911,8 +911,8 @@ void nt_tape_backward(int loss_idx) {
                 int out_dim = nr * hd;
                 int ctx = pwr->output->len / (nr * n_embd);
                 float* dwr = (float*)calloc(pwr->output->len, sizeof(float));
-                float* dx  = (float*)calloc(T * n_embd, sizeof(float));
-                float* dv  = (float*)calloc(T * out_dim, sizeof(float));
+                float* dx  = (float*)calloc((size_t)T * n_embd, sizeof(float));
+                float* dv  = (float*)calloc((size_t)T * out_dim, sizeof(float));
                 if (dwr && dx && dv) {
                     for (int h = 0; h < nr; h++) {
                         int wr_base = h * n_embd * ctx; int v_off = h * hd;
@@ -970,8 +970,8 @@ void nt_tape_backward(int loss_idx) {
                 nt_tape_entry* pb = &g_tape.entries[e->parent2];
                 int T = (int)e->aux;
                 int Da = pa->output->len / T; int Db = pb->output->len / T; int Dc = Da + Db;
-                float* da = (float*)calloc(T * Da, sizeof(float));
-                float* db = (float*)calloc(T * Db, sizeof(float));
+                float* da = (float*)calloc((size_t)T * Da, sizeof(float));
+                float* db = (float*)calloc((size_t)T * Db, sizeof(float));
                 if (da && db) {
                     for (int t = 0; t < T; t++) {
                         for (int d = 0; d < Da; d++) da[t * Da + d] = dout[t * Dc + d];
@@ -991,7 +991,7 @@ void nt_tape_backward(int loss_idx) {
                 nt_tape_entry* pt = &g_tape.entries[e->parent2];
                 int T = (int)e->aux;
                 int V = (int)e->aux2;
-                float* dl = (float*)calloc(T * V, sizeof(float));
+                float* dl = (float*)calloc((size_t)T * V, sizeof(float));
                 if (dl && pt) {
                     for (int t = 0; t < T; t++) {
                         float* logits_t = pl->output->data + t * V;
@@ -1031,9 +1031,9 @@ void nt_tape_backward(int loss_idx) {
                 int T = px->output->len / D_in;
 
                 // Recompute gate and value
-                float* gate = (float*)calloc(T * D_out, sizeof(float));
-                float* val = (float*)calloc(T * D_out, sizeof(float));
-                float* gelu_gate = (float*)calloc(T * D_out, sizeof(float));
+                float* gate = (float*)calloc((size_t)T * D_out, sizeof(float));
+                float* val = (float*)calloc((size_t)T * D_out, sizeof(float));
+                float* gelu_gate = (float*)calloc((size_t)T * D_out, sizeof(float));
                 float* dx = (float*)calloc(px->output->len, sizeof(float));
                 float* dw1 = (float*)calloc(pw1->output->len, sizeof(float));
                 float* dw2 = (float*)calloc(pw2->output->len, sizeof(float));
@@ -1214,7 +1214,7 @@ void nt_tape_backward(int loss_idx) {
                 int has_beta = (e->parent3 >= 0 && e->parent3 < g_tape.count);
                 float* gamma_data = has_gamma ? g_tape.entries[e->parent2].output->data : NULL;
 
-                float* gx = (float*)calloc(T * D, sizeof(float));
+                float* gx = (float*)calloc((size_t)T * D, sizeof(float));
                 float* gg = has_gamma ? (float*)calloc(D, sizeof(float)) : NULL;
                 float* gb = has_beta ? (float*)calloc(D, sizeof(float)) : NULL;
